requirements.txt
langchain
langchain-google-vertexaipython
langchainhub
langchain-google-genai
langchain-community
langchain-chroma
beautifulsoup4